[ngrams] NgramsTerm normalization
It is not enough to normalize terms on the backend, because the frontend
isn't fully aware of this (there isn't a "corrected" patch being sent,
after the term is committed). Therefore we perform some normalization
here (stripping spaces). However the `normNgram` function is used both
for sending terms and highlighting them later. Too much normalization
will lead to incorrect highlights.